6.2.4.p3-0113-CL173007-hwc2-fix-dest-alpha-value-in-alpha-blending
CL173007 hwc2: fix dest alpha value in alpha blending

1. fix dim layer, dim value is taken from planeAlpha but not color32.

2. fix dest alpha value

	a. When planeAlpha is not 255 (1.0f), directly disable alpha blending
          yields incorrect result. 2D hardware can not support premultiple
          with alpha for both alpha channel and color channel. Need fake a
       alpha blending with dest to get correct result.
       This is fixed.

    b. DIM optimization yields incorrect alpha result. In dim optimization,
       This is not fixed yet:
       C = Cs + Cd * (1 - As) = 0 + Cd * dim = Cd * dim
       A = Ad
       But it requires
       A = As + Ad * (1 - As) = Ad + (1 - Ad) * As
       There're still some difference.

+++ driver/driver/android/hwcomposer_v2/hwc2_blitter.h
@@ -239,7 +239,7 @@ struct blitter_layer
 
     /* Alpha blending states >>> */
     uint32_t blendFormula;  /* Layer blend formula. */
-    uint32_t globalAlpha;   /* Layer plane alpha. */
+    float globalAlpha;      /* Layer plane alpha. */
     int opaque;             /* Force no blend for tough-blend-hole opt. */
     /* <<< End alpha blending states. */
